Move code into dawe_nrm package under src
