+++ src/dawe_nrm/api/categorical_values/write.py
@@ -4,9 +4,9 @@ from rdflib import URIRef
 from rdflib.namespace import SKOS
 from rich.progress import track
 
-from src import api
-from src.graph import NRM, create_graph, serialize
-from src.schemas import LUTSchema
+from dawe_nrm import api
+from dawe_nrm.graph import NRM, create_graph, serialize
+from dawe_nrm.schemas import LUTSchema
 
 
 def write_all(path: Path, show_progress: bool = True):
