+++ tests/test_categorical_values_write_all.py
@@ -4,7 +4,7 @@ import pytest
 from rdflib import Graph, URIRef
 from rdflib.namespace import SKOS
 
-from src import api
+from dawe_nrm import api
 
 
 @pytest.fixture
